+++ src/main/lua/lrpureraw.lrdevplugin/Info.lua
@@ -18,6 +18,6 @@ return {
         id = "at.homebrew.lrpureraw", -- unique identifier for export filter
     },
 
-    VERSION = { major = 1, minor = 5, revision = 0, build = 0, },
+    VERSION = { major = 1, minor = 5, revision = 1, build = 0, },
 
 }